Add GitHub workflow to open weekly PR to update Pipfile.lock
The workflow executed by GitHub Actions will open a PR every week to update to Pipfile.lock. By using a Pipfile.lock builds as deterministic as possible and thus reproducible.

Right now the workflow is triggered every monday at 8:00 UTC and if the Pipfile is modified. Both the frequency and the timing can be adjusted by adjusting the cron: string on line 4.